Use uint64_t instead of size_t for __devicelib_assert_fail

+++ sycl/source/detail/devicelib/fallback-cassert.cpp
@@ -17,10 +17,11 @@ static const __attribute__((opencl_constant)) char assert_fmt[] =
     "Assertion `%s` failed.\n";
 
 SYCL_EXTERNAL
-extern "C" void __devicelib_assert_fail(const char *expr, const char *file,
-                                        int32_t line, const char *func,
-                                        size_t gid0, size_t gid1, size_t gid2,
-                                        size_t lid0, size_t lid1, size_t lid2) {
+extern "C" void __devicelib_assert_fail(
+    const char *expr, const char *file,
+    int32_t line, const char *func,
+    uint64_t gid0, uint64_t gid1, uint64_t gid2,
+    uint64_t lid0, uint64_t lid1, uint64_t lid2) {
   // intX_t types are used instead of `int' and `long' because the format string
   // is defined in terms of *device* types (OpenCL types): %d matches a 32 bit
   // integer, %lu matches a 64 bit unsigned integer. Host `int' and
@@ -29,8 +30,8 @@ extern "C" void __devicelib_assert_fail(const char *expr, const char *file,
       assert_fmt,
       file, (int32_t)line,
       (func) ? func : "<unknown function>",
-      (uint64_t)gid0, (uint64_t)gid1, (uint64_t)gid2,
-      (uint64_t)lid0, (uint64_t)lid1, (uint64_t)lid2,
+      gid0, gid1, gid2,
+      lid0, lid1, lid2,
       expr);
 
   // FIXME: call SPIR-V unreachable instead
